Move Page To App Router
And removed files with -static suffix
