--- pages/newyear.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-function newyear() {
-  return (
-    <div>
-      <h1>Welcome to my page</h1>
-      <button className="bg-blue-100">Click ME</button>
-    </div>
-  );
-}
-
-export default newyear;
